// lib/generated/types/model/paginatedResponseDTOMeetListItemData.ts
/**
 * Generated by orval v8.3.0 🍺
 * Do not edit manually.
 * Meeting Project Management System
 * Clean Architecture implementation with FastAPI
 * OpenAPI spec version: 1.0.0
 */
import type { MeetListItemData } from './meetListItemData';

export interface PaginatedResponseDTOMeetListItemData {
  data: MeetListItemData[];
  total: number;
  current_page: number;
  pages: number;
  is_next: boolean;
  is_prev: boolean;
  size: number;
  permissions?: string[];
  detail?: string | null;
}